Capture extra lines in the message as STDOUT
Since the Exception's message does not name the array keys, and since STDOUT (as well as STDERR) may contain multiple lines, we can't simply reference individual positions. But there are some known positions which we can take advantage of, in order to make our way to multiple lines of STDOUT.

First one is the command. The last one is the exit status, the one before exit status is cwd, and what's left is a mix of STDOUT and STDERR. For now, let's presume, that STDERR is a single line, while the rest is STDOUT.